+++ database/migrations/2025_11_21_145636_create_sanpham_table.php
@@ -11,15 +11,17 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('sanpham', function (Blueprint $table) {
-            $table->bigIncrements('masp');
-            $table->string('tensp', 30);
-            $table->integer('giasp');
-            $table->string('hinhsp', 200)->nullable();
-            $table->foreignId('madm')
-                    ->constrained('danhmuc')
-                    ->onDelete('cascade');
-        });
+        if (!Schema::hasTable('sanpham')) {
+            Schema::create('sanpham', function (Blueprint $table) {
+                $table->bigIncrements('masp');
+                $table->string('tensp', 30);
+                $table->integer('giasp');
+                $table->string('hinhsp', 200)->nullable();
+                $table->foreignId('madm')
+                        ->constrained('danhmuc', 'madm')
+                        ->onDelete('cascade');
+            });
+        }
     }
 
     /**
